requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
Pillow==9.4.0
pypng==0.20220715.0
pyTelegramBotAPI==4.10.0
qrcode==7.4.2
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
opencv-python